[Feat] 테스트 컨테이너 적용 및 숙소 상세 조회 API 구현 (#12)
* test: 테스트 컨테이너 적용

* refactor: 엔티티 수정

* feat: 숙소 상세 조회 API 작성

* test: 숙소 상세 조회 테스트 추가

* refactor: 테스트 수정

+++ backend/src/main/java/com/team3/airdnd/payment/domain/Payment.java
@@ -1,6 +1,6 @@
 package com.team3.airdnd.payment.domain;
 
-import com.team3.airdnd.accomodation.domain.Reservation;
+import com.team3.airdnd.accommodation.domain.Reservation;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -24,7 +24,7 @@ public class Payment {
     @JoinColumn(name = "payment_method_id", nullable = false)
     private PaymentMethod paymentMethod;
 
-    private Integer amount;
+    private Long amount;
 
     @Column(name = "paid_at", nullable = false)
     private LocalDateTime paidAt;
